refactor: combine interfaces to one file (interfaces.ts)
