--- src/api/ICredentials.ts
@@ -1,4 +1,0 @@
-export interface ICredentials {
-  email: string;
-  password: string;
-}
