--- src/api/IUser.ts
@@ -1,5 +1,0 @@
-export interface IUser {
-  name: string;
-  email: string;
-  password: string;
-}
