--- src/views/components/navbar/INavbar.ts
@@ -1,6 +1,0 @@
-import { IComponent } from '../../../helpers/interfaces';
-
-export interface INavbar extends IComponent {
-  setLoginState: () => void;
-  setLogoutState: () => void;
-}
